Feat: reservations CRUD and checkReserva middleware

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Api\ReservaController;
 use App\Http\Controllers\Api\TokenController;
 use App\Http\Controllers\api\UserController;
 use Illuminate\Http\Request;
@@ -13,12 +14,20 @@ use Illuminate\Support\Facades\Route;
 Route::prefix('/users')->controller(UserController::class)->group(function() {
     Route::get('/', 'index');
     Route::post('/', 'store');
-    Route::get('/{user}', 'show');
+    Route::get('/{user}', 'show')->middleware('auth:sanctum');
     Route::put('/{user}', 'update')->middleware('auth:sanctum');
-    Route::delete('/{user}', 'destroy');
+    Route::delete('/{user}', 'destroy')->middleware('auth:sanctum');
 });
 
 Route::controller(TokenController::class)->group(function() {
     Route::post('/login', 'store');
     Route::delete('/logout', 'destroy')->middleware('auth:sanctum');
+});
+
+Route::prefix('/reservas')->controller(ReservaController::class)->middleware('auth:sanctum', 'checkReserva')->group(function() {
+    Route::get('/', 'index');
+    Route::post('/', 'store');
+    Route::get('/{reserva}', 'show');
+    Route::put('/{reserva}', 'update');
+    Route::delete('/{reserva}', 'destroy');
 });
